requirements.txt
langchain
openai
faiss-cpu
tqdm
python-dotenv
